src/requirements.txt
Flask
gunicorn
requests
pythainlp
numpy
faiss-cpu
